+++ SCRIPTS/RF2-dashboards/parts/statusbar.lua
@@ -109,7 +109,7 @@ end
 M.build_ui = function(parentBox, wgt)
     local nan_w, sb_h, v_offset = tools.lcdSizeTextFixed("XXXXXXXXXX", default_font_size)
     sb_h = sb_h + 7
-    local bStatusBar = parentBox:box({x=0, y=wgt.zone.h-sb_h})
+    local bStatusBar = parentBox:box({x=0, y=wgt.zone.h-wgt.selfTopbarHeight-sb_h})
     local statusBarColor = lcd.RGB(0x0078D4)
 
     bStatusBar:rectangle({x=0, y=0,w=wgt.zone.w, h=sb_h, color=statusBarColor, filled=true})